Enable memory growth tests for wasm backend
