+++ tests/test_core.py
@@ -1710,50 +1710,11 @@ int main() {
     self.do_run(src, '''1\n1,2\n1,2,3''')
 
   def test_memorygrowth(self):
-    if self.is_wasm(): return self.skip('wasm support for memory growth in the MVP is yet unclear')
-    self.banned_js_engines = [V8_ENGINE] # stderr printing limitations in v8
-
     self.emcc_args += ['-s', 'ALLOW_MEMORY_GROWTH=0'] # start with 0
 
     # With typed arrays in particular, it is dangerous to use more memory than TOTAL_MEMORY,
     # since we then need to enlarge the heap(s).
-    src = r'''
-      #include <stdio.h>
-      #include <stdlib.h>
-      #include <string.h>
-      #include <assert.h>
-      #include "emscripten.h"
-
-      int main(int argc, char **argv)
-      {
-        char *buf1 = (char*)malloc(100);
-        char *data1 = (char*)"hello";
-        memcpy(buf1, data1, strlen(data1)+1);
-
-        float *buf2 = (float*)malloc(100);
-        float pie = 4.955;
-        memcpy(buf2, &pie, sizeof(float));
-
-        printf("*pre: %s,%.3f*\n", buf1, buf2[0]);
-
-        int totalMemory = emscripten_run_script_int("TOTAL_MEMORY");
-        char *buf3 = (char*)malloc(totalMemory+1);
-        buf3[argc] = (int)buf2;
-        if (argc % 7 == 6) printf("%d\n", (int)memcpy(buf3, buf1, argc));
-        char *buf4 = (char*)malloc(100);
-        float *buf5 = (float*)malloc(100);
-        //printf("totalMemory: %d bufs: %d,%d,%d,%d,%d\n", totalMemory, buf1, buf2, buf3, buf4, buf5);
-        assert((int)buf4 > (int)totalMemory && (int)buf5 > (int)totalMemory);
-
-        printf("*%s,%.3f*\n", buf1, buf2[0]); // the old heap data should still be there
-
-        memcpy(buf4, buf1, strlen(data1)+1);
-        memcpy(buf5, buf2, sizeof(float));
-        printf("*%s,%.3f*\n", buf4, buf5[0]); // and the new heap space should work too
-
-        return 0;
-      }
-    '''
+    src = open(path_from_root('tests', 'core', 'test_memorygrowth.c')).read()
 
     # Fail without memory growth
     self.do_run(src, 'Cannot enlarge memory arrays.')
@@ -1776,7 +1737,6 @@ int main() {
     self.emcc_args += ['--tracing']
     self.do_run(src, '*pre: hello,4.955*\n*hello,4.955*\n*hello,4.955*')
 
-  @no_wasm_backend()
   def test_memorygrowth_2(self):
     self.emcc_args += ['-s', 'ALLOW_MEMORY_GROWTH=0'] # start with 0
 
@@ -1787,47 +1747,7 @@ int main() {
 
       # With typed arrays in particular, it is dangerous to use more memory than TOTAL_MEMORY,
       # since we then need to enlarge the heap(s).
-      src = r'''
-        #include <stdio.h>
-        #include <stdlib.h>
-        #include <string.h>
-        #include <assert.h>
-        #include "emscripten.h"
-
-        int main(int argc, char **argv)
-        {
-          char *buf1 = (char*)malloc(100);
-          char *data1 = "hello";
-          memcpy(buf1, data1, strlen(data1)+1);
-
-          float *buf2 = (float*)malloc(100);
-          float pie = 4.955;
-          memcpy(buf2, &pie, sizeof(float));
-
-          printf("*pre: %s,%.3f*\n", buf1, buf2[0]);
-
-          int totalMemory = emscripten_run_script_int("TOTAL_MEMORY");
-          int chunk = 1024*1024;
-          char *buf3;
-          for (int i = 0; i < (totalMemory/chunk)+1; i++) {
-            buf3 = (char*)malloc(chunk);
-            buf3[argc] = (int)buf2;
-          }
-          if (argc % 7 == 6) printf("%d\n", memcpy(buf3, buf1, argc));
-          char *buf4 = (char*)malloc(100);
-          float *buf5 = (float*)malloc(100);
-          //printf("totalMemory: %d bufs: %d,%d,%d,%d,%d\n", totalMemory, buf1, buf2, buf3, buf4, buf5);
-          assert((int)buf4 > (int)totalMemory && (int)buf5 > (int)totalMemory);
-
-          printf("*%s,%.3f*\n", buf1, buf2[0]); // the old heap data should still be there
-
-          memcpy(buf4, buf1, strlen(data1)+1);
-          memcpy(buf5, buf2, sizeof(float));
-          printf("*%s,%.3f*\n", buf4, buf5[0]); // and the new heap space should work too
-
-          return 0;
-        }
-      '''
+      src = open(path_from_root('tests', 'core', 'test_memorygrowth_2.c')).read()
 
       # Fail without memory growth
       self.do_run(src, 'Cannot enlarge memory arrays.')
